chore: move tests under src
This is now things are organized in extensions
