+++ src/test/ExtensionInitializer.ts
@@ -9,7 +9,7 @@
 import * as Code from 'vscode';
 import * as CodeLanguageClient from 'vscode-languageclient/node';
 
-import * as Ltex from '../src/extension';
+import * as Ltex from '../extension';
 import TestTools from './TestTools';
 
 export default class ExtensionInitializer {
