+++ tsconfig.json
@@ -8,6 +8,7 @@
     "moduleResolution": "node",
     "outDir": "out",
     "resolveJsonModule": true,
+    "rootDir": "src",
     "skipLibCheck": true,
     "sourceMap": true,
     "strict": true,
